Tidy up API of pd crate
