+++ tikv-client-pd/src/lib.rs
@@ -1,7 +1,8 @@
 // Copyright 2018 TiKV Project Authors. Licensed under Apache-2.0.
 
+#[doc(inline)]
 pub use cluster::{Cluster, Connection};
-use kvproto::pdpb;
+#[doc(inline)]
 pub use tikv_client_common::{security::SecurityManager, Error, Result};
 
 #[macro_use]
@@ -9,31 +10,3 @@ extern crate log;
 
 mod cluster;
 mod timestamp;
-
-trait PdResponse {
-    fn header(&self) -> &pdpb::ResponseHeader;
-}
-
-impl PdResponse for pdpb::GetStoreResponse {
-    fn header(&self) -> &pdpb::ResponseHeader {
-        self.get_header()
-    }
-}
-
-impl PdResponse for pdpb::GetRegionResponse {
-    fn header(&self) -> &pdpb::ResponseHeader {
-        self.get_header()
-    }
-}
-
-impl PdResponse for pdpb::GetAllStoresResponse {
-    fn header(&self) -> &pdpb::ResponseHeader {
-        self.get_header()
-    }
-}
-
-impl PdResponse for pdpb::UpdateGcSafePointResponse {
-    fn header(&self) -> &pdpb::ResponseHeader {
-        self.get_header()
-    }
-}
